requirements.txt
asgiref==3.2.10
backcall==0.2.0
blessings==1.7
boto3==1.15.16
botocore==1.18.17
click==7.1.2
colorama==0.4.4
coverage==5.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-configurations==2.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-model-utils==4.0.0
django-nose==1.4.7
django-rest-passwordreset==1.1.0
django-storages==1.10.1
django-unique-upload==0.2.1
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
flake8==3.8.4
future==0.18.2
gunicorn==20.0.4
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mock==4.0.2
newrelic==5.20.1.150
nltk==3.5
nose==1.3.7
nose-progressive==1.5.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
social-auth-app-django
social-auth-core
django-pg-copy
django-sendgrid-v5
django-extensions
requests